Layout views that are hidden.
iOS sizeThatFits: and layoutSubviews still work on views that are hidden. Reasons you might want to lay out a hidden view:
 1. Because it's going to animate to be visible later and needs to be in the right place
 2. Because it's actually a drawable and doesn't depend on the hidden property

If we did want to re-add this, we'd at very least need to fix it for right align and bottom align. For right align, for example, we should return
    CGRectMake(frame.size.width, 0, frame.size.width, frame.size.height);
Otherwise subsequent views that make use of the origin property of the hidden view might get messed up.